Restructured, added src directory for proper GOPATH parsing for config package.
Implemented a basic config file, in JSON format.
